Feat: Updates Character and Item models
With this update the Character and Item models are updated to use JSON
data persistance, Helper functions are implemented in the Helper package
to assist in reading the JSON files.
A Google library for converting Java Objects to and from Json is implemented
for converting Character and Item objects to and from JSON. The implementation
of how to do this was based on the lab work we did in class.